requirements.txt
streamlit>=1.30.0
openai>=1.0.0
python-dotenv
Pillow
pdf2image
pandas
numpy